sdasd

+++ Project Codes/TMD/TMDWeb/BBLocationRead.aspx.cs
@@ -114,8 +114,7 @@ namespace targeted_marketing_display
         {
             if (e.CommandName == "BillboardUpdateInfo")
             {
-
-                int index = Convert.ToInt32(e.CommandArgument);
+               
 
 
                 // Retrieve the row that contains the link button from the Rows collection.
@@ -123,9 +122,7 @@ namespace targeted_marketing_display
                 LinkButton lbButton1 = sender as LinkButton;
                 GridViewRow gvRow1 = (GridViewRow)lbButton1.NamingContainer;
 
-                //Need to Retrieve userID to edit user
-                Label lb_BillboardID = (Label)gvRow1.FindControl("lb_BillboardID");
-
+               
                 Session["BillboardID"] = lb_BillboardID.Text;
 
                 Response.Redirect("BBLocationUpdate.aspx");
@@ -136,6 +133,17 @@ namespace targeted_marketing_display
         {
             if (e.CommandName == "DeleteBBMessage")
             {
+                SqlConnection conn = null;
+                SqlDataReader reader = null;
+
+
+
+                // instantiate and open connection
+                conn = new
+                    SqlConnection(@"Data Source=L33527\CHEEEFANGSQL;Initial Catalog=Targeted_Marketing_Display;Persist Security Info=True;User ID=root;Password=passw8rd");
+                conn.Open();
+             
+
                 int index = Convert.ToInt32(e.CommandArgument);
 
 
@@ -154,16 +162,43 @@ namespace targeted_marketing_display
 
                 bObj = bDao.getBillboardByID(lb_msgId.Text);
                 string  BBCode= bObj.BillboardCode;
+                SqlCommand cmdcount = new SqlCommand("Select count(*) as total from BillboardLocation as b inner join AdvertisementLocation al on b.BillboardID=al.BillboardID where al.BillboardID=@ID", conn);
+                SqlParameter paramCount = new SqlParameter();
+                paramCount.ParameterName = "@ID";
+                paramCount.Value = bObj.BillboardID.ToString();
+                cmdcount.Parameters.Add(paramCount);
+                SqlDataAdapter sdacount = new SqlDataAdapter();
+                DataTable dt = new DataTable();
+                cmdcount.Connection = conn;
+                sdacount.SelectCommand = cmdcount;
+                sdacount.Fill(dt);
+                for(int i = 0;i< dt.Rows.Count; i++)
+                {
+                    int count = Convert.ToInt32(dt.Rows[i]["total"]);
+                    if (count == 0)
+                    {
+                        Boolean insCnt = bDao.deleteBillboard(lb_msgId.Text);
+
+                        //VIC: never inform if the delete is successful or not?
+                        alertSuccess.Visible = true;
+                        DeleteFailure.Visible = false;
+                        msgSuccess.Text = " Billboard #" + BBCode + " Has Been Deleted Successfully!";
+
+                    }
+                    else
+                    {
+                        DeleteFailure.Visible = true;
+                        alertSuccess.Visible = false;
+                        LabelError.Text = "There are " + count + "Advertisement(s) in Billboard " + BBCode + ".Please Delete existing advertisement(s) first before proceeding to delete the Billboard.";
+                    }
+                }
 
 
 
 
-                Boolean insCnt = bDao.deleteBillboard(lb_msgId.Text);
 
-                //VIC: never inform if the delete is successful or not?
-                alertSuccess.Visible = true;
-                msgSuccess.Text =" Billboard #"+ BBCode + " Has Been Deleted Successfully!";
 
+                
                 Database db = new Database();
 
                 SqlCommand cmd = new SqlCommand("Select * from [BillboardLocation] WHERE Status = 1");
